[bugfix] move to hash history to support static site
The demo site is deployed on Github Page which can only serve as a
static site. If we use browser native history as React Router history
backend, it would return 404 if requesting for any subdirectory such as
`/heroes/` or `/heroes/1`. Thus, the compatible solution is moving the
history backend to traditional hash history.